+++ table_parser/src/main.cpp
@@ -20,8 +20,8 @@ struct my_data {
     sub_data d;
 };
 
-bool ParseSubDataCallback(const char* s, unsigned len, void* data,
-                          unsigned size, void* context) {
+bool ParseSubDataCallback(const char* s, size_t len, void* data,
+                          size_t size, void* context) {
     sub_data* p = reinterpret_cast<sub_data*>(data);
 
     if (size != sizeof(sub_data)) return false;
@@ -56,11 +56,11 @@ const char* my_data_text =
 int main() {
     vector<my_data> results;
     vector<string> errors;
-    unsigned count = tp::parse_all(my_data_text, my_data_desc, results, errors);
+    size_t count = tp::parse_all(my_data_text, my_data_desc, results, errors);
 
-    for (unsigned i = 0; i < errors.size(); ++i) cout << errors[i] << endl;
-    for (unsigned i = 0; i < results.size(); ++i) {
-        for (unsigned j = 0; j < results[i].count_a; ++j)
+    for (int i = 0; i < errors.size(); ++i) cout << errors[i] << endl;
+    for (int i = 0; i < results.size(); ++i) {
+        for (int j = 0; j < results[i].count_a; ++j)
             cout << results[i].a[j] << " ";
         cout << results[i].b << " ";
         cout << results[i].c << " ";
