+++ table_parser/src/table_parser.cpp
@@ -12,8 +12,8 @@
 namespace tp {
 
 // 支持针对10进制带符号32位整数的解析
-static bool parse_int_callback(const char *s, unsigned len, void *data,
-                               unsigned size, void *context) {
+static bool parse_int_callback(const char *s, size_t len, void *data,
+                               size_t size, void *context) {
     UNUSED(context);
 
     if (size != sizeof(int)) {
@@ -37,7 +37,7 @@ static bool parse_int_callback(const char *s, unsigned len, void *data,
 
     // 数字
     int ret = 0;
-    for (unsigned i = 0; i < len; ++i) {
+    for (int i = 0; i < len; ++i) {
         char c = s[i];
         if (c >= '0' && c <= '9') {
             ret = ret * 10 + (c - '0');
@@ -56,8 +56,8 @@ static bool parse_int_callback(const char *s, unsigned len, void *data,
 
 // 支持符合浮点数表达的解析
 // 参见json.org的语法描述
-static bool parse_float_callback(const char *s, unsigned len, void *data,
-                                 unsigned size, void *context) {
+static bool parse_float_callback(const char *s, size_t len, void *data,
+                                 size_t size, void *context) {
     UNUSED(context);
 
     if (size != sizeof(float)) {
@@ -85,8 +85,8 @@ static bool parse_float_callback(const char *s, unsigned len, void *data,
     bool neg_power = false;
     float power = 0;
 
-    unsigned state = 0;
-    for (unsigned i = 0; i < len; ++i) {
+    int state = 0;
+    for (int i = 0; i < len; ++i) {
         char c = s[i];
         switch (state) {
             case 0:
@@ -169,8 +169,8 @@ static bool parse_float_callback(const char *s, unsigned len, void *data,
 }
 
 // 支持字符串格式
-static bool parse_string_callback(const char *s, unsigned len, void *data,
-                                  unsigned size, void *context) {
+static bool parse_string_callback(const char *s, size_t len, void *data,
+                                  size_t size, void *context) {
     UNUSED(context);
 
     if (size < len + 1) {
@@ -184,7 +184,6 @@ static bool parse_string_callback(const char *s, unsigned len, void *data,
 
 TableParser::TableParser(const char *src, const ColumnDescriptor desc[])
     : _src(src), _desc(desc), _line(1) {
-    // std::strcpy(_err, "ok");
     std::snprintf(_err, sizeof(_err), "%s", "ok");
 }
 
@@ -192,7 +191,6 @@ TableParser::TableParser(const TableParser &org) {
     _src = org._src;
     _desc = org._desc;
     _line = org._line;
-    // std::strcpy(_err, org._err);
     std::snprintf(_err, sizeof(_err), "%s", org._err);
 }
 
@@ -200,7 +198,6 @@ TableParser &TableParser::operator=(const TableParser &rhs) {
     _src = rhs._src;
     _desc = rhs._desc;
     _line = rhs._line;
-    // std::strcpy(_err, rhs._err);
     std::snprintf(_err, sizeof(_err), "%s", rhs._err);
     return *this;
 }
@@ -216,7 +213,7 @@ TableParser &TableParser::operator=(const TableParser &rhs) {
  *   针对非array字段，element构成元素中不得出现'\t'
  *   针对array字段，element构成元素中不得出现','和'\t'
  */
-ParseResult TableParser::parse(void *p, unsigned size) {
+ParseResult TableParser::parse(void *p, size_t size) {
     char c = *_src;
     if (c == '\0') {
         return KEOF;
@@ -256,7 +253,7 @@ ParseResult TableParser::parse(void *p, unsigned size) {
         // 解析一个元素
         if (!parse_end && desc.is_array) {
             char *end_of_count = nullptr;
-            unsigned count = std::strtoul(_src, &end_of_count, 10);
+            size_t count = std::strtoul(_src, &end_of_count, 10);
 
             if (end_of_count == _src) {
                 ret = KERROR;
@@ -285,7 +282,7 @@ ParseResult TableParser::parse(void *p, unsigned size) {
 
             if (!parse_end) {
                 // 设置数组大小描述内存
-                *reinterpret_cast<unsigned *>(
+                *reinterpret_cast<size_t *>(
                     (void *)((char *)p + desc.array_counter_offset)) = count;
 
                 // 依次解析数组元素
@@ -424,7 +421,7 @@ ParseResult TableParser::parse(void *p, unsigned size) {
 const char *TableParser::last_error() const { return _err; }
 
 ParseResult TableParser::parse_element(unsigned idx, const char *s,
-                                       unsigned len, void *data) {
+                                       size_t len, void *data) {
     const ColumnDescriptor &desc = _desc[idx];
     parser_callback cb;
